Check in two changes: The first change (which is disabled) compactifies all of the function constant pools into the global constant pool, in an attempt to reduce the amount of duplication and overhead.  Unfortunately, as the comment indicates, this is not yet a win, so it is disabled.
The second change sorts the typeid's so that those types that can be used
by instructions in the program appear earlier in the table than those that
cannot (such as structures and arrays).  This causes the instructions to
be able to use the dense encoding more often, saving about 5K on 254.gap.
This is only a .65% savings though, unfortunately. :(

llvm-svn=10754