Added "PrinterHelper" interface (include/AST/PrinterHelper) that can be passed as an (optional) argument to StmtPrinter to customize printing of AST nodes.
Used new PrinterHelper interface to enhance printing and visualization
of CFGs.  The CFGs now illustrate the semantic connectives between
statements and terminators, wheras in the previous printing certain
expressions would (visible) be printed multiple times to reflect which
expressions used the results of other expressions.

The end result is that the CFG is easier to read for flow of
expression values (following principles similar to the LLVM IR).

llvm-svn=41651