Create FunctionType::isValidArgumentType to go along with isValidReturnType. Also create isValidElementType for ArrayType, PointerType, StructType and VectorType.
Make LLParser use them. This closes up some holes like an assertion failure on:

  %x = type {label}

but largely doesn't change any semantics. The only thing we accept now which
we didn't before is vectors of opaque type such as "<4 x opaque>". The opaque
can be resolved to an int or float when linking.

llvm-svn=73016